+++ rust/src/api/application.rs
@@ -15,7 +15,7 @@ pub struct ApplicationListOptions {
     pub exclude_apps_with_svix_play_endpoints: Option<bool>,
 
     /// Limit the number of returned items
-    pub limit: Option<i32>,
+    pub limit: Option<u64>,
 
     /// The iterator returned from a prior invocation
     pub iterator: Option<String>,
