+++ svix-cli/src/cmds/api/connector.rs
@@ -8,13 +8,16 @@ use svix::models::*;
 pub struct ConnectorListOptions {
     /// Limit the number of returned items
     #[arg(long)]
-    pub limit: Option<i32>,
+    pub limit: Option<u64>,
+
     /// The iterator returned from a prior invocation
     #[arg(long)]
     pub iterator: Option<String>,
+
     /// The sorting order of the returned items
     #[arg(long)]
     pub order: Option<Ordering>,
+
     #[arg(long)]
     pub product_type: Option<ConnectorProduct>,
 }
