+++ svix-cli/src/cmds/api/operational_webhook_endpoint.rs
@@ -8,10 +8,12 @@ use svix::models::*;
 pub struct OperationalWebhookEndpointListOptions {
     /// Limit the number of returned items
     #[arg(long)]
-    pub limit: Option<i32>,
+    pub limit: Option<u64>,
+
     /// The iterator returned from a prior invocation
     #[arg(long)]
     pub iterator: Option<String>,
+
     /// The sorting order of the returned items
     #[arg(long)]
     pub order: Option<Ordering>,
