+++ svix-cli/src/cmds/api/streaming_event_type.rs
@@ -8,13 +8,16 @@ use svix::models::*;
 pub struct StreamingEventTypeListOptions {
     /// Limit the number of returned items
     #[arg(long)]
-    pub limit: Option<i32>,
+    pub limit: Option<u64>,
+
     /// The iterator returned from a prior invocation
     #[arg(long)]
     pub iterator: Option<String>,
+
     /// The sorting order of the returned items
     #[arg(long)]
     pub order: Option<Ordering>,
+
     /// Include archived (deleted but not expunged) items in the response.
     #[arg(long)]
     pub include_archived: Option<bool>,
